+++ .gitignore
@@ -18,8 +18,6 @@ DerivedData
 *.hmap
 *.ipa
 
-
-
 # Bundler
 .bundle
 
@@ -39,4 +37,5 @@ Carthage/Build
 # Note: if you ignore the Pods directory, make sure to uncomment
 # `pod install` in .travis.yml
 #
-# Pods/
+Pods/
+*.xcworkspace
