+++ Example/Podfile
@@ -1,22 +1,13 @@
 source 'https://cdn.cocoapods.org/'
 
-
 use_frameworks!
 
 platform :ios, '15.0'
 
 target 'TikiClient_Example' do
-  pod 'TikiClient', :path => '../'
-  
-  pod 'TikiSdkDebug', '3.0.0', :configurations => 'Debug'
-  pod 'TikiSdkRelease', '3.0.0', :configurations => 'Release'
-  
-  pod 'AppAuth'
-  pod 'CryptoSwift', '~> 1.8.1'
+  pod 'TikiClient', :path => '../../'
 
   target 'TikiClient_Tests' do
     inherit! :search_paths
-
-    
   end
 end
