--- TikiClient/Classes/Configuration.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-public struct Configuration{
-    let terms: String
-    var clientId: String
-    var clientSecret: String?
-}
