+++ TikiClient/Classes/TikiError.swift
