+++ TikiClient/Classes/Ui/Retailer/RetailerEnum.swift
