+++ TikiClient/Classes/Ui/Util/Color.swift
