+++ TikiClient/Classes/Ui/Util/Component/BottomSheetHeader.swift
