+++ TikiClient/Classes/Ui/Util/Component/CircularProgressView.swift
