+++ TikiClient/Classes/Ui/Util/Component/ScreenHeader.swift
