+++ TikiClient/Classes/Ui/Util/RoundedCorner.swift
