+++ TikiClient/Classes/Ui/Util/TikiImages.swift
