+++ TikiClient/Classes/Ui/Util/Window.swift
