--- TikiClient/Classes/Util/Extension/Color.swift
@@ -1,46 +1,0 @@
-/*
- * Copyright (c) TIKI Inc.
- * MIT license. See LICENSE file in the root directory.
- */
-
-import SwiftUI
-
-public extension Color {
-    
-    static let tikiBlack: Color = Color(hex: "FF000000")
-    
-    static let tikiDarkGray:Color =  Color(hex: "99000000")
-    
-    static let tikiLightGray:Color = Color(hex: "15000000")
-    
-    static let tikiGreen: Color = Color(hex: "FF00B272")
-    
-    static let tikiRed: Color = Color(hex: "C73000")
-    
-    static let tikiWhite: Color = Color(hex: "FFFFFFFF")
-    
-    init(hex: String) {
-        let hex = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
-        var int: UInt64 = 0
-        Scanner(string: hex).scanHexInt64(&int)
-        let a, r, g, b: UInt64
-        switch hex.count {
-        case 3: // RGB (12-bit)
-            (a, r, g, b) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
-        case 6: // RGB (24-bit)
-            (a, r, g, b) = (255, int >> 16, int >> 8 & 0xFF, int & 0xFF)
-        case 8: // ARGB (32-bit)
-            (a, r, g, b) = (int >> 24, int >> 16 & 0xFF, int >> 8 & 0xFF, int & 0xFF)
-        default:
-            (a, r, g, b) = (1, 1, 1, 0)
-        }
-
-        self.init(
-            .sRGB,
-            red: Double(r) / 255,
-            green: Double(g) / 255,
-            blue:  Double(b) / 255,
-            opacity: Double(a) / 255
-        )
-    }
-}
