--- TikiClient/Classes/Util/Extension/DataString.swift
@@ -1,33 +1,0 @@
-//
-//  DataString.swift
-//  TikiClient
-//
-//  Created by Jesse Monteiro Ferreira on 19/02/24.
-//
-
-import Foundation
-
-extension String {
-    func base64Decoded() -> String? {
-        var st = self
-            .replacingOccurrences(of: "_", with: "/")
-            .replacingOccurrences(of: "-", with: "+")
-        let remainder = self.count % 4
-        if remainder > 0 {
-            st = self.padding(toLength: self.count + 4 - remainder,
-                              withPad: "=",
-                              startingAt: 0)
-        }
-        guard let d = Data(base64Encoded: st, options: .ignoreUnknownCharacters) else{
-            return nil
-        }
-        return String(data: d, encoding: .utf8)
-    }
-    var imageFromBase64: UIImage? {
-        guard let imageData = Data(base64Encoded: self, options: .ignoreUnknownCharacters) else {
-            return nil
-        }
-        return UIImage(data: imageData)
-    }
-    
-}
