+++ TikiClient/Classes/Util/String.swift
@@ -1,0 +1,19 @@
+//
+//  UrlSafe.swift
+//  TikiClient
+//
+//  Created by Jesse Monteiro Ferreira on 26/02/24.
+//
+
+import Foundation
+
+extension String {
+
+    var imageFromBase64: UIImage? {
+       
+    }
+    
+}
+
+
+
